refactor: rename files to be equal to their use_case_functions
