+++ src/features/document/use_cases/remove_use_case.py
